requirements.txt
arrow==1.2.2
binaryornot==0.4.4
briefcase==0.3.7
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cookiecutter==1.7.3
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
jinja2-time==0.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
pexpect==4.8.0
pip==22.1.1
poyo==0.5.0
ptyprocess==0.7.0
python-dateutil==2.8.2
python-slugify==6.1.2
requests==2.27.1
setuptools==62.3.2
six==1.16.0
smmap==5.0.0
text-unidecode==1.3
tomli==2.0.1
urllib3==1.26.9
wheel==0.37.1
PyYAML==6.0
schema==0.7.5
Zenity==2.0.0